+++ admin/install.php
@@ -140,8 +140,8 @@ function runInstall($data) {
 		<meta charset="UTF-8">
 		<meta name="viewport" content="width=device-width, initial-scale=1.0">
 		<meta name="robots" content="noindex, nofollow">
-		<link href="<?php echo ADMIN_STYLES; ?>/install.css" rel="stylesheet">
-		<link href="<?php echo STYLES; ?>/buttons.css" rel="stylesheet">
+		<link href="<?php echo ADMIN_STYLES.'/install.min.css'; ?>" rel="stylesheet">
+		<link href="<?php echo STYLES.'/button.min.css'; ?>" rel="stylesheet">
 	</head>
 	<body>
 		<h1>ReallySimpleCMS</h1>